+++ NEWS.md
@@ -1,7 +1,7 @@
 # snakecase 0.10.9000
 
 * changed title case to build up on sentence case instead of parsed_case.
-* abbreviations are no case-insensitive and consistently always turned into upper case for title-, upper-camel-, lower-camel- and mixed-case.
+* abbreviations are no matched case-insensitive and for title-, upper-camel-, lower-camel- and mixed case the supplied input specifies the formatting of the output. 
 
 # snakecase 0.10.0 (16.05.19)